+++ module/StepStar/src/StepStar/Controller/Generate/GenerateController.php
@@ -3,7 +3,9 @@
 namespace StepStar\Controller\Generate;
 
 use Application\Controller\AbstractController;
+use Fichier\FileUtils;
 use Laminas\Http\Response;
+use Laminas\Session\Container as SessionContainer;
 use StepStar\Facade\Generate\GenerateFacadeAwareTrait;
 use StepStar\Form\Generate\GenerateForm;
 use StepStar\Service\Fetch\FetchServiceAwareTrait;
@@ -17,6 +19,12 @@ class GenerateController extends AbstractController
     use LogServiceAwareTrait;
 
     private GenerateForm $generateForm;
+    private SessionContainer $sessionContainer;
+
+    public function __construct()
+    {
+        $this->sessionContainer = new SessionContainer(__CLASS__);
+    }
 
     public function setGenerateForm(GenerateForm $generateForm): void
     {
@@ -44,15 +52,18 @@ class GenerateController extends AbstractController
                     $this->flashMessenger()->addErrorMessage("Aucune these trouvee avec les criteres specifies");
                 }
 
-                $logs = $this->generateFacade->generateFilesForTheses($theses, $command);
+                $log = $this->generateFacade->generateFilesForTheses($theses, $command);
 
-                /** @var \StepStar\Entity\Db\Log $log */
-                foreach ($logs as $log) {
-                    if ($log->isSuccess()) {
-                        $this->flashMessenger()->addSuccessMessage($log->getLogToHtml());
-                    } else {
-                        $this->flashMessenger()->addErrorMessage($log->getLogToHtml());
-                    }
+                // pour l'action de téléchargement ci-après, on met à disposition dans la session le chemin du fichier généré
+                $resulFilePath = $log->getZipFilePath() ?: $log->getTefFilePath();
+                $this->sessionContainer->offsetSet('resulFilePath', $resulFilePath);
+
+                if ($log->isSuccess()) {
+                    $downloadUrl = $this->url()->fromRoute('step-star/generation/telecharger-tef');
+                    $this->flashMessenger()->addSuccessMessage($log->getLogToHtml() .
+                        "<br><a href='$downloadUrl'>Télécharger le ou les fichiers TEF générés</a>");
+                } else {
+                    $this->flashMessenger()->addErrorMessage($log->getLogToHtml());
                 }
 
                 return $this->redirect()->refresh();
@@ -61,6 +72,27 @@ class GenerateController extends AbstractController
 
         return [
             'form' => $this->generateForm,
+            'downloadEnabled' => $this->sessionContainer->offsetExists('resulFilePath'),
         ];
+    }
+
+    private function isDownloadResultEnabled(): bool
+    {
+        return $this->sessionContainer->offsetExists('resulFilePath');
+    }
+
+    /**
+     * Action pour envoyer au client web le fichier TEF généré ou un zip en cas de fichiers TEF multiples.
+     * Le chemin du fichier sur le serveur est transmis à cette action via la session.
+     */
+    public function telechargerTefAction(): Response
+    {
+        if (!$this->sessionContainer->offsetExists('resulFilePath')) {
+            return $this->redirect()->toRoute('step-star/generation');
+        }
+
+        $resulFilePath = $this->sessionContainer->offsetGet('resulFilePath');
+
+        FileUtils::downloadFile($resulFilePath);
     }
 }