+++ module/StepStar/src/StepStar/Service/Zip/ZipServiceAwareTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace StepStar\Service\Zip;
+
+trait ZipServiceAwareTrait
+{
+    protected ZipService $zipService;
+
+    public function setZipService(ZipService $zipService): void
+    {
+        $this->zipService = $zipService;
+    }
+}